IfcGeomIterator: Add a products_to_include set
Enables filtering of products, alongside the existing
names_to_include_or_exclude and entities_to_include_or_exclude filters.

The added set does not follow the --include or --exclude commands, it
only includes. The argument for doing this is to enabling filtering of a
large sub-set of the IFC, e.g. a building or a storey. --include and
--exclude will then only affect the selection within this subset.

products_to_include being empty is interpreted as "no sub-set
selected", i.e. select entire IFC.