adjust attribute type comparison character and characterization
Ifc Selector does not work in comparisons with numbers and boolean properties, I propose to mischaracterize the attribute type in string and leave this attribution up to the user, using quotation marks in the case of string attributes. in addition, I included 'f.comparison' in the query because it is always using the equals comparator.