BBIM remove Blender Manifest file
BlenderBIM is not working yet for Blender 4.2. Keeping the manifest around means
the add-on would try to be installed (by Blender) as an extension. Removing the add-on
makes it to install correctly (as legacy add-on).

This should give the developers more time to decide how to tackle the conversion to extension.

Note: blender_version_min should have been 4.2 since extensions were not existent before that,
and this range is only looked at in relation to extensions.