Bonsai offers to install Git and ifcmerge
If Git isn't installed on Windows, the Git panel now offers to install
it from the Microsoft Store using `winget`.

Then if ifcmerge isn't installed on Linux or Windows, the Git panel
offers to install it from copies bundled in the Bonsai ZIP package.

Note that ifcmerge.exe increases the size of the Bonsai ZIP download by
about 7MB.