Fix 3x slower major speed regression introduced in bc3a1ee4346e2dc19ff0583bcb76b4f4c1a073b2
The commit that introduced the problem was trying to optimise rendering
dense meshes with the opening decorator. However, it introduced more
calculation in the initial project load and stored more data on every
object. This resulted in models loading 3x slower for me which is pretty
major. The original problem needs further investigation.